This commit adds support for measuring hardware counters for nested loops. Consider the following loop structure:
L1
--statements
--L2 (nested)
----statements
----L3
------statements

In the previous version, measuring L1 and L2 at the same time was not allowed.
With this commit, that is possible. This allows for example to measure inclusive
hardware events for L1 (which includes counts for all computations within L1), and
also get a separate counts for L2 and L3 in a single run.

There are other changes:
1. Changed the hardware counter specification to: PEBIL_HWC0, PEBIL_HWC1 ...
from just HWC0, HWC1.

2. Changed the set identification env variable to PEBIL_HWC_SET_NUMBER from
HWC_SET_NUMBER.

3. Added a PEBIL_DISABLE_NESTING env variable that disables nested measurements
(which is enabled by default). If this variable is set to 1 or higher, the tool
checks if the loops specified for measurements are nested. If they are nested,
the tool produces an error and exits out.