This commit provides needed changes in the FunctionTimer tool to differentiate functions in cases where there are multiple functions defined in the binary with the same name.
This was observed for SWFFT application, where as the issue
report states, there were multiple definitions for "apply"
function.

To get around the issues, we now also report the address for
the head block of each function.

Changes need to be made on the PerfPal side to ensure we
take this new information into account when displaying
function hotspots and when inlining stats (the latter
might already be in place).