Solved Issue #68, which saw a static analysis error with the instruction `vshuff32x4`. This was being represented as a 16 x 128 vector instruction, which implies 2048 bits instead of the maximum 512.
Deciding how to actually represent this instruction was tough because,
according to the intel manual, the instruction operates at the 128-bit
granularity (hence the 128 bit element size), but the writemask updates
the destination operand with a granularity of 32-bit elements. So, an
argument could be made for both `4 x 128` and `16 x 32`.

We chose to use the latter because this is consistent with how we treat
similar instructions, like vextractf32x4.