Solves Issue #66:
During instrumentation, a number of functions are hidden for any of a
number of reasons. We used to print this to standard out with a
cryptic message. Now, we print a more understandable message to a file
called `<APP>.<tool>.hiddenFunctions`. The file includes the function
name, the hash function, and why it was hidden.