Fixed Issue #98.
Push/pop instructions usually do an implicit memory
operation. However, when they push/pop to a memory address, they do an
additional explicit memory operation. The problem from this issue was
that the explicit memory address was loaded twice in the latter case
during Address Stream Intercept. This is because our assembler for
loading the effective address from a memory operation did not make this
distinction for the case of push/pop instructions.

This commit adds that distinction into the assembler.