This commit fixes two errors, both of which involved a 0 being recorded as an address during cache simulation.
The first bug was with regard to the following instruction:
`mov %fs:0x28, %rax`. Unfortunately, due to some deprecated code, we were
only loading the value of `%fs` into the address, instead of calculating the
address at `%fs:0x28`. The FS segment refers to thread local storage, but for
some reason, the value of `%fs` is zero (even if we do the instruction
`lea %fs:0x28, address`). After some gdb, it became clear that after
`mov %fs:0x28, %rax`, the value in `%rax` is the thread id plus `0x28`. Thus,
the fix for emitting the address collection for an instruction that uses the
FS segment of the form `mov %fs:[ADDR], %reg` is to:

```
lea [ADDR], %rcx    (where %rcx is our temporary storage register)
add %fs:0x10, %rcx  (%fs:0x10 holds the thread id)
```

The second bug a minor one for the following instruction:
`rep stosq %rax, %es:(%rdi)`. In this instruction, the operands are implied,
not explicit. It is only a store (di), but our instruction classifer had it
as a load and a store (dsi). After making that change and another minor
bug in emitAddressComputation, the instrumented binary worked just fine.