Reversed a change made by Issue #144. This change was made because of a comment in types.h (in the disassembler), which said that the byte position of the offset/imm/modrm began with 1 (instead of 0). After a lot of review, I believe that this comment was written (by me!) erroneously. I do not think it refers to the modrm byte at all, and instead refers to where the offset or immediate should be.
The confusing part is when you have an offset of 0. Sometimes, that
offset is not encoded in the instruction, and is instead implied.
We still set the position, but this position will be greater than
the number of bytes in the instruction. I have gone back to our
disassembler tests and added clarifying comments there.

I have also fixed the comment in types.h.