Fixed Issue #157.
The Cray C/C++ compilers produce a binary that is organized very
differently than the GNU format, which broke a number of hardcoded
assumptions in PEBIL. The following assumptions were generalized:

* PEBIL extended the binary to the front, starting with the first
text section (keeping that first text section in place). The cray
compiler actually generates a binary with .rodata before the first
text section. We do not want this to move, so now it extends the
binary up to the first text or data section.
* PEBIL assumed that the first text section was ".init". When
looking for the first text section, this now can be .init or
.rodata (see above). This should be generalized better in the
future.
* Since a change to the GNU hash table reorders the symbol table,
the gnu hash table must be changed before the sysv table. PEBIL
assumed that the GNU hash table would be first. This commit forces
the GNU table to be updated before the sysv hash table.
* PEBIL made assumptions about which sections come after the dynamic
symbol table. It correctly shifted sections after adding an entry
to the table, but it hardcoded which dynamic table entries needed
to be updated (when a section move, the corresponding table entry
must be updated). This commit generalizes this code: it checks if
the section comes after the symbol table. If so, it updates the
corresponding table entry.