Added METASIM_DS_SIZE
Added the reading of environment variable METASIM_DS_SIZE.
There is an issue with adding to DataStructure after we have already set the
size of the stream stats. Added the option to manually set this if the user is
having the issue. If no METASIM_DS_SIZE is set, will use the default values
that were calculated.