Large bug with ProcessAllBuffers: The buffer position is updated at the block level. When we fill out a buffer entry, we would get the new position and work backwards. With ProcessAllBuffers, another thread could reset the buffer position (to 0), and then the filled out entry would overwrite at best the 0th buffer entry and at worst who knows what.
This PR changes that. To change the least amount of assembly logic
possible, we went with an idea put forth by Emmet: Save the
position before incrementing it. The use the old position and work
forwards. This guarantees that the 0th position is never overwritten.

This does not guarantee that every address will be collected. With
a lot of threads, some will be lost. We are willing to accept this
for now.