Refactored MemoryStreamHandler::Process
MemoryStreamHandler::Process use to be passed a BufferEntry that
was pretty specific to x86 architecture in how it dealt with vector Entries.
That has been refactored to only pass the necessary info to the underlying
function calls so that they can be invoked without needing to know what a
BufferEntry is.