Issue #172: A warning from the ReuseDistance library was causing deadlock (this seems to happen with threads + fprintf + locks). Solved the issue by just removing the warning. Talked to Emmet about why he introduced the warning, and it really boiled down to "sampling could cause issues" which we already warn the user about.
I also saw another place where there is a potential for deadlock.
It was theoretically possible for Thread 1 to grab the data
structure module read lock at the beginning of ProcessThreadBuffer,
for Thread 2 to grab the sampler write lock and then attempt to
grab the data structure module write lock in ProcessAllBuffers, and
then for Thread 1 to try to use the sampler lock. Reordered the
locks in ProcessAllBuffers to try and avoid this.