Fixed a bug caused by Emmet's refactor of the Address Stream Library and Allyson's Data Centric branch. We actually predicted this would be an issue and placed a TODO to fix before merging the two branches but we missed it.
The bug was that the refactor unravels a buffer entry and passes all
addresses to each handler. To save performance, it created a single
dynamically allocated array for the Driver and passed the addresses
down. This does not work with threads since the threads share the
Driver.

The fix was a simple one. To keep the optimization of only allocating
a single array for each thread, we allocate the array with the
AddressStreamStats object and store it there.