Issue #202: Fixed bug #202. This was a bug with instrumenting an aocc-generated binary.
AOCC appears to generate binaries with an empty data section that
PEBIL did not know how to handle. The fix is to just call it a
raw section instead of a data section since we don't need an empty
section.

A second bug was caused by the fact that the .init_array section was
placed just above the .dynamic section. In __libc_csu_init, the code
goes through the .init_array section and stops when it hits the
address for the following section. AOCC puts the .dynamic section
after it, but we move the dynamic section. That reference to the
.dynamic section gets updated and the __libc_csu_init code becomes
wrong. The current fix is to just not update anchors for
__libc_csu_init, which is okay since we don't instrument this
function. But it may require a more general solution if the
wedging algorith requires the update.