Issue #202: Updated configure and Make to be able to use nvhpc. Configuration may require CC=mpicc etc (wrapping nvc).
As of August 2024, the nvhcp compilers could not use a "builtin"
command that the configure script uses to detect the presence of
the AVX512F instruction set. We agreed to assume that gcc would
be available, so the configure script now switches to it before
checking for AVX512F.

The nvhpc compilers do not have the `-dynamic` linker option. So,
removing these for now to see if they are even necessary.